pyproject.toml
[project]
name = "shinyliveapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "shinylive>=0.7.3",
    "uvicorn>=0.34.0",
]
